% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_words.R
\name{to_words}
\alias{to_words}
\title{From integers to spanish spelled quantities.}
\usage{
to_words(x)
}
\arguments{
\item{x}{An integer up to 1e22 or a string of numbers up to 60 digits}
}
\value{
A string for the same integer number in spanish.
}
\description{
Takes any number up to 1e22 or string of numbers
up to 60 digits and returns spanish characters.
The general strategy is to follow the "natural way"
of reading numbers in Spanish: divide them in
groups of 3 from right to left, read them in
hundreds, using mapply, (that's the work of the
function (convert_3_digits())) and attach the suffix
as appropriate (thousands, millions, etc.)
with mapply.
Spanish uses long scale, and that is the one
used in this code (1 billion = 1e12). See
(https://en.wikipedia.org/wiki/Long_and_short_scales)
}
\examples{
# Converting numbers (can be negative)
to_words(31254865)
# converting strings of digits (can have a "-" as the first character)
to_words("-542687982546720143")
}
\keyword{integers,}
\keyword{money,}
\keyword{quantities.}
\keyword{spanish,}
